requirements.txt
azureml-sdk
azure-mgmt-web
numpy
pandas
scikit-learn
streamlit
requests


